% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RG199R2.R
\name{calc_cf_weld}
\alias{calc_cf_weld}
\title{Chemistry Factor (CF) Calculation for Weld Metal}
\usage{
calc_cf_weld(Cu, Ni)
}
\arguments{
\item{Cu}{numeric vector, copper content in weight percent (wt\%). Must be between 0 and 0.4.}

\item{Ni}{numeric vector, nickel content in weight percent (wt\%). Must be between 0 and 1.2.}
}
\value{
A numeric vector representing the chemistry factor (CF) in degrees Fahrenheit (°F).
}
\description{
Computes the chemistry factor (CF) for weld metals based on the tables provided in Regulatory Guide 1.99 Rev. 2.
The CF is determined by interpolating values from the provided table based on input copper (Cu) and nickel (Ni) content.
}
\examples{
# Example: Calculate CF for weld metals with given Cu and Ni values
rpvetc:::calc_cf_weld(0.1, 0.6)

}
