% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{conv}
\alias{conv}
\title{Unit Conversion Function Registry}
\format{
An object of class \code{list} of length 8.
}
\usage{
conv
}
\description{
A list of commonly used unit conversion functions. Each element is named by its conversion direction.
}
\examples{
conv$F_to_C(212)       # 100
conv$ksi_to_MPa(10)    # 68.9476
conv$ftlb_to_J(20)     # 27.1164

}
\keyword{datasets}
