% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CR3391.R
\name{CR3391}
\alias{CR3391}
\title{CR3391}
\usage{
CR3391(
  product_form,
  Cu,
  Ni,
  fluence = 0,
  output = c("TTS", "CF", "FF", "SD"),
  temperature_unit = c("Celsius", "Fahrenheit")
)
}
\arguments{
\item{product_form}{character vector c("B", "F", "P", "W")}

\item{Cu}{numeric vector, wt\%}

\item{Ni}{numeric vector, wt\%}

\item{fluence}{nummeric vector n/cm2}

\item{output}{string c("TTS", "CF", "FF", "SD")}

\item{temperature_unit}{character vector, c("Celsius", "Fahrenheit")}
}
\value{
TTS as given temperature_unit
}
\description{
Calculate TTS using NUREG/CR3391 vol.2 report by Guthrie (1983).
}
\examples{
CR3391("B", 0.1, 0.6, 1e19)
}
