% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E900_15.R, R/E900_flux.R
\name{E900_15_SD}
\alias{E900_15_SD}
\title{E900_15_SD}
\usage{
E900_15_SD(product_form, TTS)

E900_15_SD(product_form, TTS)
}
\arguments{
\item{product_form}{character vector c("F", "P", "W")}

\item{TTS}{numeric vector, degC}
}
\value{
SD of TTS (degC)

SD of TTS (degC)
}
\description{
Calculate SD of ASTM E900-15, vectorized.

Calculate SD of ASTM E900-15, vectorized.
}
