% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E900_15.R, R/E900_flux.R
\name{E900_15_TTS}
\alias{E900_15_TTS}
\title{E900_15_TTS}
\usage{
E900_15_TTS(product_form, Cu, Ni, Mn, P, temperature, fluence)

E900_15_TTS(product_form, Cu, Ni, Mn, P, temperature, fluence)
}
\arguments{
\item{product_form}{character vector c("F", "P", "W")}

\item{Cu}{numeric vector, wt\%}

\item{Ni}{numeric vector, wt\%}

\item{Mn}{numeric vector, wt\%}

\item{P}{numeric vector, wt\%}

\item{temperature}{numeric vector, temperature_unit}

\item{fluence}{numeric vector, n/cm2}
}
\value{
TTS (degC)

TTS (degC)
}
\description{
Calculate TTS of ASTM E900-15, vectorized.

Calculate TTS of ASTM E900-15, vectorized.
}
