% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E900_15.R
\name{E900_15_TTS1}
\alias{E900_15_TTS1}
\title{E900_15_TTS1}
\usage{
E900_15_TTS1(product_form, Ni, Mn, P, temperature, fluence)
}
\arguments{
\item{product_form}{character vector c("F", "P", "W")}

\item{Ni}{numeric vector, wt\%}

\item{Mn}{numeric vector, wt\%}

\item{P}{numeric vector, wt\%}

\item{temperature}{numeric vector, temperature_unit}

\item{fluence}{numeric vector, n/cm2}
}
\value{
TTS (degC)
}
\description{
Calculate TTS1 of ASTM E900-15, vectorized.
}
