% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E900_flux.R
\name{E900_flux_TTS2}
\alias{E900_flux_TTS2}
\title{E900_flux_TTS2}
\usage{
E900_flux_TTS2(product_form, Cu, Ni, P, temperature, fluence, flux)
}
\arguments{
\item{product_form}{character vector c("F", "P", "W")}

\item{Cu}{numeric vector, wt\%}

\item{Ni}{numeric vector, wt\%}

\item{P}{numeric vector, wt\%}

\item{temperature}{numeric vector, temperature_unit}

\item{fluence}{numeric vector, n/cm2}

\item{flux}{numeric vector, n/cm2/s}
}
\value{
TTS2 (degC)
}
\description{
Calculate TTS2 of ASTM E900-15.
}
