% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expand_vectors}
\alias{expand_vectors}
\title{Expand input vectors to the maximum common length}
\usage{
expand_vectors(...)
}
\arguments{
\item{...}{A list of vectors to be expanded.}
}
\value{
A list of vectors, all of the same length.
}
\description{
Ensures all input vectors are of the same length by replicating elements
of length 1 to match the maximum non-zero length among the inputs.
}
