% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RG199R2.R
\name{linear_interpolate_2d}
\alias{linear_interpolate_2d}
\title{linear_interpolate_2d}
\usage{
linear_interpolate_2d(x_values, y_values, table, x, y)
}
\arguments{
\item{x_values, }{numeric vector, x축 값들}

\item{y_values, }{numeric vector, y축 값들}

\item{table, }{numeric matrix, 2D 테이블 (행: y축, 열: x축)}

\item{x, }{numeric, 보간하려는 x 지점}

\item{y, }{numeric, 보간하려는 y 지점}
}
\value{
interpoated value
}
\description{
Provide interpolated value from 2D table
}
