% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP3319.R
\name{NP3319}
\alias{NP3319}
\title{NP3319}
\usage{
NP3319(
  product_form,
  Cu,
  Ni,
  fluence = 0,
  output = c("TTS", "CF", "FF"),
  temperature_unit = c("Celsius", "Fahrenheit")
)
}
\arguments{
\item{product_form}{character vector c("B", "F", "P", "W")}

\item{Cu}{numeric vector, wt\%}

\item{Ni}{numeric vector, wt\%}

\item{fluence}{numeric vector n/cm2}

\item{output}{string c("TTS", "CF", "FF")}

\item{temperature_unit}{character vector, c("Celsius", "Fahrenheit")}
}
\value{
output as given unit
}
\description{
Calculate TTS using NP3319(EPRI NP-3319 TTS model by Odette, 1984).
}
\examples{
NP3319("B", 0.1, 0.6, 1e19)
}
