% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RG199R2.R
\name{rg199_ff}
\alias{rg199_ff}
\title{Fluence Factor (FF)
Computes Fluence Factor from fluence in n/cm² (numeric vector)}
\usage{
rg199_ff(fluence)
}
\description{
Fluence Factor (FF)
Computes Fluence Factor from fluence in n/cm² (numeric vector)
}
