% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RG199R2.R
\name{rg199_margin}
\alias{rg199_margin}
\title{Marging from CF and fluence, SD
Generalized computation: TTS = CF × FF}
\usage{
rg199_margin(cf, fluence, sd)
}
\description{
Marging from CF and fluence, SD
Generalized computation: TTS = CF × FF
}
