% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RG199R2.R
\name{RG199R2_CF}
\alias{RG199R2_CF}
\title{RG199R2_CF}
\usage{
RG199R2_CF(
  product_form = NULL,
  Cu = NULL,
  Ni = NULL,
  SV_flu = NULL,
  SV_tts = NULL,
  CF = NULL
)
}
\arguments{
\item{product_form}{character vector, c("B", F", "P", "W")}

\item{Cu}{numeric vector, wt\%}

\item{Ni}{numeric vector, wt\%}

\item{SV_flu}{numeric vector, n/cm2}

\item{SV_tts}{numeric vector, degF}

\item{CF}{numeric vector, degF}
}
\value{
CF as degF
}
\description{
Provide chemistry factor
}
