% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RG199R2.R
\name{rg199r2_CF_base}
\alias{rg199r2_CF_base}
\title{rg199r2_CF_base}
\usage{
rg199r2_CF_base(Cu, Ni)
}
\arguments{
\item{Cu}{numeric vector, wt\%}

\item{Ni}{numeric vector, wt\%}
}
\value{
CF as degF
}
\description{
Provide CF of base from the tables of Regulatory Guide 1.99 Rev. 2.
}
