% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RG199R2.R
\name{RG199R2_CF_weld}
\alias{RG199R2_CF_weld}
\title{RG199R2_CF_weld}
\usage{
RG199R2_CF_weld(Cu, Ni)
}
\arguments{
\item{Cu}{numeric vector, wt\%}

\item{Ni}{numeric vector, wt\%}
}
\value{
CF as degF
}
\description{
Provide CF of weld from the tables of Regulatory Guide 1.99 Rev. 2.
}
