% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RG199R2.R
\name{RG199R2_P1}
\alias{RG199R2_P1}
\title{Regulatory Guide 1.99 Rev. 2 Position 1.1 Table-based Embrittlement Property Calculator}
\usage{
RG199R2_P1(
  product_form,
  Cu = NULL,
  Ni = NULL,
  output = c("CF", "SD"),
  output_unit = c("Celsius", "Fahrenheit"),
  use_names = FALSE
)
}
\description{
Computes radiation embrittlement-related properties based on the U.S. NRC Regulatory Guide 1.99 Rev. 2 (1988),
specifically Position 1.1. This model uses tabulated Chemistry Factors and fixed Standard Deviations
based on material composition and neutron fluence.
}
