% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RG199R2.R
\name{RG199R2_SV}
\alias{RG199R2_SV}
\title{RG199R2_SV}
\usage{
RG199R2_SV(
  fluence,
  SV_tts,
  output = c("TTS", "CF", "SD"),
  temperature_unit = c("Celcius", "Fahrenheit")
)
}
\arguments{
\item{fluence}{numeric vector, n/cm2}

\item{SV_tts}{numeric vector, TTS from surveillance tests}

\item{output}{string c("TTS", "CF", "SD)}

\item{temperature_unit}{string c("Celcius", "Fahrenheit")}
}
\value{
A numeric vector of CF, TTS, or SD (depending on \code{output}).
}
\description{
Provide TTS or CF of Regulatory Guide 1.99 Rev. 2
}
\examples{
RG199R2_SV(c(0, 0.1, 0.2) * 1e19, c(0, 10, 20), output = "TTS", temperature_unit = "F")
}
