% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CR3391.R
\name{CR3391}
\alias{CR3391}
\title{CR3391}
\usage{
CR3391(
  product_form = NULL,
  Cu = NULL,
  Ni = NULL,
  fluence = NULL,
  output = c("TTS", "CF", "FF", "SD"),
  temperature_unit = c("Celsius", "Fahrenheit")
)
}
\arguments{
\item{product_form}{character vector. One of c("B", "F", "P", "W").}

\item{Cu}{numeric vector, wt\% (0~100)}

\item{Ni}{numeric vector, wt\% (0~100)}

\item{fluence}{numeric vector n/cm2 (0 이상)}

\item{output}{character c("TTS", "CF", "FF", "SD")}

\item{temperature_unit}{character c("Celsius", "Fahrenheit")}
}
\value{
A numeric vector corresponding to \code{output} in the specified \code{temperature_unit}.
}
\description{
Provide TTS, CF, FF, or SD of Guthrie's model in NUREG/CR3391 (1983).
}
\examples{
# 1) TTS 계산 (모든 인자 필요)
CR3391(product_form = "B", Cu = 0.1, Ni = 0.6, fluence = 1e19)

# 2) FF만 계산 (Cu, Ni 불필요)
CR3391(product_form = "F", fluence = 1e19, output = "FF")

}
