% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CR3391.R
\name{CR3391}
\alias{CR3391}
\title{CR3391}
\usage{
CR3391(form, Cu, Ni, fl)
}
\arguments{
\item{Cu}{number \emph{wt\%}.}

\item{Ni}{number weight-percentage}

\item{product}{string c("F", "P", "W")}

\item{Mn}{number weight-percentage}

\item{P}{number weight-percentage}

\item{temperature}{number degree Celcius}

\item{fluence}{number n/cm2}

\item{temperature_unit}{string c("degC", "degF")}

\item{fluence_area}{string c("cm2", "m2")}

\item{output}{string c("tts", "sd", "tts1", "tts2)}
}
\value{
TTS as given temperature_unit
}
\description{
Calculate TTS using CR3391 (NUREG/CR3391 vol.2 by Guthrie (1983).
}
\examples{
E900_15("P", 0.2, 0.18, 1.36, 0.012, 290, 2.56894e18) # should be 31.743721
}
