% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E900_15.R
\name{E900_15}
\alias{E900_15}
\title{ASTM E900-15e2}
\usage{
E900_15(
  product_form,
  Cu,
  Ni,
  Mn,
  P,
  temperature,
  fluence,
  fluence_area = "cm2",
  temperature_unit = "Celsius",
  output = "TTS"
)
}
\arguments{
\item{product_form}{character vector c("F", "P", "W")}

\item{Cu}{numeric vector, wt\%}

\item{Ni}{numeric vector, wt\%}

\item{Mn}{numeric vector, wt\%}

\item{P}{numeric vector, wt\%}

\item{temperature}{numeric vector, temperature_unit}

\item{fluence}{numeric vector, neutrons/fluence_area}

\item{fluence_area}{string c("cm2", "m2") (default: "cm2")}

\item{temperature_unit}{string c("Celsius", "Fahrenheit") (default: "Celsius")}

\item{output}{character c("TTS", "SD", "TTS1", "TTS2") (default: "TTS")}
}
\value{
TTS or SD as given temperature_unit
}
\description{
Provide TTS or SD of ASTM E900-15e2.
}
\examples{
E900_15("P", 0.2, 0.18, 1.36, 0.012, 290, 2.56894e18) # should be 31.74387
}
