% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E900_15.R
\name{E900_15}
\alias{E900_15}
\title{ASTM E900-15e2}
\usage{
E900_15(
  product,
  Cu,
  Ni,
  Mn,
  P,
  temperature,
  fluence,
  temperature_unit = "degC",
  fluence_area = "cm2",
  output = "tts"
)
}
\arguments{
\item{product}{string c("F", "P", "W")}

\item{Cu}{number \emph{wt\%}.}

\item{Ni}{number weight-percentage}

\item{Mn}{number weight-percentage}

\item{P}{number weight-percentage}

\item{temperature}{number degree Celcius}

\item{fluence}{number n/cm2}

\item{temperature_unit}{string c("degC", "degF")}

\item{fluence_area}{string c("cm2", "m2")}

\item{output}{string c("tts", "sd", "tts1", "tts2)}
}
\value{
TTS as given temperature_unit
}
\description{
Provide TTS  or SD of ASTM E900-15e2.
}
\examples{
E900_15("P", 0.2, 0.18, 1.36, 0.012, 290, 2.56894e18)  # should be 31.743721
}
