% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expand_vectors}
\alias{expand_vectors}
\title{입력 벡터들을 가장 긴 길이에 맞게 확장하는 함수}
\usage{
expand_vectors(...)
}
\arguments{
\item{...}{확장할 벡터들의 리스트}
}
\value{
길이가 확장된 벡터들의 리스트
}
\description{
입력 벡터들을 가장 긴 길이에 맞게 확장하는 함수
}
