% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP3319.R
\name{NP3319}
\alias{NP3319}
\title{NP3319}
\usage{
NP3319(
  product_form = NULL,
  Cu = NULL,
  Ni = NULL,
  fluence = NULL,
  output = c("TTS", "CF", "FF", "SD"),
  temperature_unit = c("Celsius", "Fahrenheit")
)
}
\arguments{
\item{product_form}{character vector c("B","F","P","W")}

\item{Cu}{numeric vector, wt\% (0~100). Not required if only \code{output="FF"}.}

\item{Ni}{numeric vector, wt\% (0~100). Not required if only \code{output="FF"}.}

\item{fluence}{numeric vector (n/cm2). Required unless you set \code{output="CF"} only.}

\item{output}{character c("TTS", "CF", "FF", "SD")}

\item{temperature_unit}{character c("Celsius", "Fahrenheit")}
}
\value{
Numeric vector of CF, FF, or TTS in the specified unit.
}
\description{
Provide TTS, CF, FF, or SD of Odette's model in EPRI NP-3319 (1984).
}
\examples{
# 1) TTS 계산 (모든 인자 필요)
NP3319(product_form = "B", Cu = 0.1, Ni = 0.6, fluence = 1e19)

# 2) FF만 계산 (Cu, Ni 불필요)
NP3319(product_form = "F", fluence = 1e19, output = "FF")

}
