% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RG199R2.R
\name{RG199R2}
\alias{RG199R2}
\title{RG199R2}
\usage{
RG199R2(
  product_form = NULL,
  Cu = NULL,
  Ni = NULL,
  fluence = NULL,
  tts = NULL,
  output = c("TTS", "CF", "sigma"),
  temperature_unit = c("Celcius", "Fahrenheit")
)
}
\arguments{
\item{product_form}{character vector, c("B", F", "P", "W")}

\item{Cu}{numeric vector, wt\%}

\item{Ni}{numeric vector, wt\%}

\item{fluence}{numeric vector, n/cm2}

\item{tts}{numeric vector, TTS from surveillance tests}

\item{output}{string c("TTS", "CF", "sigma)}

\item{temperature_unit}{string c("Celcius", "Fahrenheit")}
}
\value{
TTS or CF as given condition
}
\description{
Provide TTS or CF of Regulatory Guide 1.99 Rev. 2
}
\examples{
RG199R2("B", 0.2, 0.18, 2.56894e18) # should be 31.74387
}
