% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RG199R2.R
\name{RG199R2}
\alias{RG199R2}
\title{RG199R2}
\usage{
RG199R2(
  product_form = NULL,
  Cu = NULL,
  Ni = NULL,
  fluence = NULL,
  SV_flu = NULL,
  SV_tts = NULL,
  CF = NULL,
  output = c("TTS", "CF", "FF", "SD"),
  temperature_unit = c("Celsius", "Fahrenheit"),
  verbose = FALSE
)
}
\arguments{
\item{product_form}{character vector, c("B", F", "P", "W")}

\item{Cu}{numeric vector, wt\%}

\item{Ni}{numeric vector, wt\%}

\item{fluence}{numeric vector, n/cm2}

\item{SV_flu}{numeric vector, n/cm2}

\item{SV_tts}{numeric vector, temperature unit}

\item{CF}{numeric vector, temperature unit}

\item{output}{character c("TTS", "CF", "FF", "SD")}

\item{temperature_unit}{character c("Celsius", "Fahrenheit")}

\item{verbose}{logical TRUE or FALSE}
}
\value{
various value according to output
}
\description{
Provide TTS or CF of Regulatory Guide 1.99 Rev. 2 tables
}
\examples{
RG199R2("B", 0.2, 0.18, 2.56894e18) # should be 31.74387
}
