% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RG199R2.R
\name{RG199R2_CF_by_SV}
\alias{RG199R2_CF_by_SV}
\title{RG199R2_CF_by_SV}
\usage{
RG199R2_CF_by_SV(SV_flu, SV_tts)
}
\arguments{
\item{SV_flu}{numeric vector, n/cm2}

\item{SV_tts}{numeric vector, degF}
}
\value{
CF as degF
}
\description{
Provide CF calculated using TTS of the surveillance test results
}
