% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RG199R2.R
\name{RG199R2_CF_table}
\alias{RG199R2_CF_table}
\title{RG199R2_CF_table}
\usage{
RG199R2_CF_table(product_form, Cu, Ni)
}
\arguments{
\item{product_form}{character vector, c("B", "W")}

\item{Cu}{numeric vector, wt\%}

\item{Ni}{numeric vector, wt\%}
}
\value{
CF as degF
}
\description{
Provide CF from the tables of Regulatory Guide 1.99 Rev. 2
}
