% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RG199R2.R
\name{RG199R2_P2}
\alias{RG199R2_P2}
\title{Regulatory Guide 1.99 Rev. 2 Position 2.1 Surveillance-Based Embrittlement Property Calculator}
\usage{
RG199R2_P2(
  SV_flu,
  SV_tts,
  product_form = NULL,
  output = c("CF", "SD"),
  output_unit = c("Celsius", "Fahrenheit"),
  SV_tts_unit = c("Celsius", "Fahrenheit")
)
}
\description{
Computes embrittlement-related properties using surveillance test data based on the interpretation
of U.S. NRC Regulatory Guide 1.99 Rev. 2 (1988), Position 2.1.
}
\details{
The function estimates key embrittlement parameters such as Transition Temperature Shift (TTS),
Chemistry Factor (CF), Fluence Factor (FF), Standard Deviation (SD), and Regulatory Margin using
surveillance capsule data from reactor pressure vessel materials.
}
