% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RG199R2.R
\name{RG199R2_TTS}
\alias{RG199R2_TTS}
\title{RG199R2_TTS}
\usage{
RG199R2_TTS(CF, fluence)
}
\arguments{
\item{CF}{numeric vector, degF}

\item{fluence}{numeric vector, n/cm2}
}
\value{
output TTS as degF
}
\description{
Provide TTS from CF and fluence
}
