% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_baseweld}
\alias{to_baseweld}
\title{Product Form을 "B"(Base) 또는 "W"(Weld)로 표준화하는 함수}
\usage{
to_baseweld(product_form)
}
\arguments{
\item{product_form}{문자열 벡터 ("B", "F", "P", "W")}
}
\value{
"B" 또는 "W"로 변환된 문자열 벡터
}
\description{
Product Form을 "B"(Base) 또는 "W"(Weld)로 표준화하는 함수
}
